requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
distro==1.8.0
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==4.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
oauthlib==3.2.2
openai==1.3.5
Pillow==10.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.2
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
